+++ apps/api/src/bootstrap.ts
@@ -20,7 +20,7 @@ import { validateEnv } from './config/env-validator';
 
 import * as packageJson from '../package.json';
 import { setupSwagger } from './app/shared/framework/swagger/swagger.controller';
-import { HttpRequestHeaderKeysEnum } from './app/shared/framework/types';
+import { corsOptionsDelegate } from './config/cors';
 
 const extendedBodySizeRoutes = ['/v1/events', '/v1/notification-templates', '/v1/workflows', '/v1/layouts'];
 
@@ -108,32 +108,4 @@ export async function bootstrap(expressApp?): Promise<INestApplication> {
   Logger.log(`Started application in NODE_ENV=${process.env.NODE_ENV} on port ${process.env.PORT}`);
 
   return app;
-}
-
-const corsOptionsDelegate: Parameters<INestApplication['enableCors']>[0] = function (req, callback) {
-  const corsOptions: Parameters<typeof callback>[1] = {
-    origin: false as boolean | string | string[],
-    preflightContinue: false,
-    maxAge: 86400,
-    allowedHeaders: Object.values(HttpRequestHeaderKeysEnum),
-    methods: ['GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'PATCH', 'OPTIONS'],
-  };
-
-  if (['test', 'local'].includes(process.env.NODE_ENV) || isWidgetRoute(req.url) || isBlueprintRoute(req.url)) {
-    corsOptions.origin = '*';
-  } else {
-    corsOptions.origin = [process.env.FRONT_BASE_URL];
-    if (process.env.WIDGET_BASE_URL) {
-      corsOptions.origin.push(process.env.WIDGET_BASE_URL);
-    }
-  }
-  callback(null as unknown as Error, corsOptions);
-};
-
-function isWidgetRoute(url: string) {
-  return url.startsWith('/v1/widgets');
-}
-
-function isBlueprintRoute(url: string) {
-  return url.startsWith('/v1/blueprints');
 }
